compose-configs/optional-associated-runtimes/airflow-marquez/requirements.txt
apache-airflow-providers-openlineage
apache-airflow-providers-postgres
confluent-kafka
openlineage-python[kafka]
pyegeria
apache-airflow-providers-apache-kafka
asgiref
unitycatalog
